Split UUID data type out into separate Cabal package
- Moved UUID data type definition to Data.UUID.Types (along with the
  accompanying Builder) in new "uuid-types" package.

- Previous Data.UUID.Internal modules moved to
  Data.UUID.Types.Internal and exported from "uuid-types" so that the
  "uuid" package has full access to internals.

- Data.UUID.Types is reexported as Data.UUID in the "uuid" package,
  which hopefully should preserve API-compatibility.

- Tests have been split into "uuid-types"-specific tests and and
  "uuid"-specific tests. Both sets of tests run during Travis builds.

- Basic benchmarks have been moved to "uuid-types", but all the
  "generate a UUID" benchmarks were moved to "uuid" since that's where
  the generation algorithms live.

- Dependencies in each of the two packages have been reduced as much
  as possible.
