+++ uuid-types/CHANGES
@@ -1,0 +1,3 @@
+1.0.0
+
+- Initial split from "uuid-1.3.8"
