+++ uuid-types/CONTRIBUTORS
@@ -10,3 +10,4 @@ Sergei Trofimovich
 davean
 Francesco Mazzoli
 Michael Snoyman
+Bardur Arantsson
