+++ uuid-types/uuid-types.cabal
@@ -1,0 +1,89 @@
+Name: uuid-types
+Version: 1.0.0
+Copyright: (c) 2008-2014 Antoine Latter
+Author: Antoine Latter
+Maintainer: aslatter@gmail.com
+License: BSD3
+License-file: LICENSE
+
+Category: Data
+Build-Type: Simple
+Cabal-Version: >= 1.8
+
+Description:
+ This library contains type definitions for Universally Unique Identifiers
+ and basic conversion functions.
+
+ See <http://en.wikipedia.org/wiki/UUID> for the general idea.
+
+Synopsis: Type definitions for Universally Unique Identifiers
+
+Homepage: https://github.com/aslatter/uuid
+Bug-Reports: https://github.com/aslatter/uuid/issues
+
+Extra-Source-Files:
+     CHANGES
+     CONTRIBUTORS
+
+
+Library
+ Build-Depends: base >=3 && < 5,
+                binary >= 0.4 && < 0.8,
+                bytestring >= 0.9 && < 1.1,
+                deepseq >= 1.3 && < 1.5,
+                hashable (>= 1.1.1.0 && < 1.2.0) || (>= 1.2.1 && < 1.3),
+                random >= 1.0.1 && < 1.2
+
+ Exposed-Modules:
+   Data.UUID.Types
+   -- Exposed for companion projects; *NOT* part of the official API:
+   Data.UUID.Types.Internal.Builder
+   Data.UUID.Types.Internal
+
+ Extensions: DeriveDataTypeable
+ Ghc-Options:        -Wall
+
+source-repository   head
+  type:             git
+  location:         https://github.com/aslatter/uuid.git
+
+Test-Suite testuuid
+    Type:              exitcode-stdio-1.0
+    Main-is:           TestUUID.hs
+    Hs-source-dirs:    tests
+    Extensions:        DeriveDataTypeable
+    Ghc-Options:       -Wall -fno-warn-orphans
+    Build-Depends:     base >= 3 && < 5,
+                       uuid-types,
+                       bytestring >= 0.9 && < 1.1,
+                       HUnit >=1.2 && < 1.3,
+                       QuickCheck >=2.4 && < 2.8,
+                       test-framework == 0.8.*,
+                       test-framework-hunit == 0.3.*,
+                       test-framework-quickcheck2 == 0.3.*
+
+benchmark benchmark
+    Type:              exitcode-stdio-1.0
+    Main-is:           BenchUUID.hs
+    Hs-source-dirs:    tests
+    Extensions:        DeriveDataTypeable, CPP
+    Ghc-Options:       -Wall -fno-warn-orphans
+    Build-depends:     base == 4.*,
+                       uuid-types,
+                       bytestring >= 0.9 && < 1.1,
+                       containers >= 0.4 && < 0.6,
+                       criterion >= 0.4 && < 0.9,
+                       deepseq >= 1.1 && < 1.4,
+                       random >= 1.0.1 && < 1.2
+
+benchmark benchbytestring
+  Type:              exitcode-stdio-1.0
+  Hs-Source-Dirs:    tests
+  Main-Is:           BenchByteString.hs
+  Ghc-Options:       -threaded -rtsopts -Wall -O2 -fno-warn-orphans
+  Build-Depends:     base                 >= 4       && < 5
+                   , uuid-types
+                   , bytestring           >= 0.10
+                   , criterion            >= 0.8
+                   , random               >= 1.0
+                   , QuickCheck           >= 2
