+++ uuid/CONTRIBUTORS
@@ -1,0 +1,13 @@
+In order of appearance:
+
+Antoine Latter
+Jason Dusek
+Tim Newsham
+Mark Lentczner
+Neil Mitchell
+Bas van Dijk
+Sergei Trofimovich
+davean
+Francesco Mazzoli
+Michael Snoyman
+Bardur Arantsson
