+++ uuid/Setup.hs
@@ -1,0 +1,4 @@
+import Distribution.Simple
+main :: IO ()
+main = defaultMain
+
